examples/raspberry-pi/requirements.txt
paho-mqtt>=1.6,<2.0
psutil>=5.9
